// src/app/interfaces/order.interface.ts
export interface OrderInterface {
  address: string;
  author: string;
  centerNumArray: number[];
  confirm: boolean;
  orderDate: string;
  description: string;
  distance: number;
  executor: string;
  geometry: number;
  mode: string;
  payType: string;
  photo: string[];
  price: number;
  region: number;
  seenOrder: string[];
  subscription: boolean;
  type: string;
  written: string[];
  role: string;
}